style(refactoring) : devided styled component code merge into one file, styled components devide into style file
