--- src/styles/Home/PresentRanking/StyledPresentRankingItemBrandName.ts
@@ -1,6 +1,0 @@
-import styled from '@emotion/styled';
-
-export const StyledPresentRankingItemBrandName = styled.p`
-  color: ${({ theme }) => theme.sementicPalette.textDisabled};
-  padding: ${({ theme }) => theme.spacing.spacing1};
-`;
