--- src/styles/Home/PresentRanking/StyledPresentRankingItemDiv.ts
@@ -1,6 +1,0 @@
-import styled from '@emotion/styled';
-
-export const StyledPresentRankingItemDiv = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
