--- src/styles/Home/PresentRanking/StyledPresentRankingItemImage.ts
@@ -1,6 +1,0 @@
-import styled from '@emotion/styled';
-
-export const StyledPresentRankingItemImage = styled.img`
-  position: relative;
-  width: 100%;
-`;
