--- src/styles/Home/PresentRanking/StyledPresentRankingItemPrasentPrice.ts
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const StyledPresentRankingItemPrasentPrice = styled.p`
-  color: ${({ theme }) => theme.typography.body2Bold};
-  padding-left: ${({ theme }) => theme.spacing.spacing1};
-  padding-top: ${({ theme }) => theme.spacing.spacing2};
-`;
