--- src/styles/Home/PresentRanking/StyledPresentRankingItemPresentItem.ts
@@ -1,6 +1,0 @@
-import styled from '@emotion/styled';
-
-export const StyledPresentRankingItemPresentItem = styled.p`
-  color: ${({ theme }) => theme.typography.body2Regular};
-  padding-left: ${({ theme }) => theme.spacing.spacing1};
-`;
