--- src/styles/Home/PresentRanking/StyledPresentRankingNumContainer.ts
@@ -1,17 +1,0 @@
-import styled from '@emotion/styled';
-
-export const StyledPresentRankingNumContainer = styled.div<{ index: number }>`
-  position: absolute;
-  background-color: ${({ index, theme }) =>
-    index <= 3 ? theme.palette.red600 : theme.palette.gray600};
-  width: 20px;
-  height: 20px;
-  z-index: 10;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 4px;
-  border-radius: 2px;
-  color: ${({ theme }) => theme.palette.blue00};
-  ${({ theme }) => theme.typography.label2Bold}
-`;
