+++ src/styles/Home/RankingTagItem/StyledRankingSexTagItemBtn.tsx
@@ -7,14 +7,17 @@ const StyledRankingSexTagItemBtn = styled.button<SexTagBtnProps>`
   background-color: ${({ theme }) => theme.sementicPalette.backgroundDefault};
   border: none;
   div {
-    background-color: ${({ theme, isSelected }) => (isSelected ? theme.palette.blue700 : theme.palette.blue200)};
+    background-color: ${({ theme, isSelected }) =>
+      isSelected ? theme.palette.blue700 : theme.palette.blue200};
     border-radius: 10px;
     width: 45px;
     height: 45px;
   }
   p {
-    color: ${({ theme, isSelected }) => (isSelected ? theme.palette.blue700 : theme.palette.gray500)};
-    ${({ theme, isSelected }) => (isSelected ? theme.typography.label1Bold : theme.typography.label1Regular)}
+    color: ${({ theme, isSelected }) =>
+      isSelected ? theme.palette.blue700 : theme.palette.gray500};
+    ${({ theme, isSelected }) =>
+      isSelected ? theme.typography.label1Bold : theme.typography.label1Regular}
   }
 `;
