--- src/styles/Login/StyledLoginComponentContainerDiv.tsx
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const StyledLoginComponentContainerDiv = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
