--- src/styles/Login/StyledLoginKakoLogo.tsx
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const StyledLoginKakoLogo = styled.div`
-  ${({ theme }) => theme.typography.title2Bold}
-  margin: ${({ theme }) => theme.spacing.spacing11};
-  font-size: 30px;
-`;
