--- src/styles/Order/OrderContainer/StyledRecipientsModalContainer.tsx
@@ -1,10 +1,0 @@
-import styled from '@emotion/styled';
-
-export const StyledRecipientsModalContainer = styled.div`
-  width: 100%;
-  margin: 20px 0px 20px 0px;
-  padding: 6px 0px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-`;
