--- src/styles/Order/OrderContainer/SyltedOrderInput.tsx
@@ -1,9 +1,0 @@
-import styled from '@emotion/styled';
-
-export const SyltedOrderInput = styled.input`
-  padding: 6px 12px;
-  width: 95%;
-  &:focus {
-    outline: none;
-  }
-`;
